Seperate the models to different apps
