+++ backend/api/settings.py
@@ -47,6 +47,8 @@ THIRD_PARTY_APPS = [
 LOCAL_APPS = [
     "core",
     "users",
+    "listings",
+    "orders",
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
