+++ backend/users/models.py
@@ -42,65 +42,9 @@ class Address(BaseModel):
         return f"{self.street}, {self.city}, {self.state} {self.zip_code}"
 
 
-class Listing(BaseModel):
-    title = models.CharField(max_length=255)
-    image = models.ImageField(upload_to="listing_images/")
-    description = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveIntegerField()
-    owner = models.ForeignKey(User, on_delete=models.CASCADE)
-    active = models.BooleanField(default=True)
-    category = models.ForeignKey("Category", on_delete=models.CASCADE)
-
-    class Meta:
-        verbose_name = "Listing"
-        verbose_name_plural = "Listings"
-
-    def __str__(self):
-        return self.title
-
-
-class Category(BaseModel):
-    name = models.CharField(max_length=255)
-    description = models.TextField()
-
-    class Meta:
-        verbose_name = "Category"
-        verbose_name_plural = "Categories"
-
-    def __str__(self):
-        return self.name
-
-
-class Coupon(BaseModel):
-    code = models.CharField(max_length=255)
-    discount = models.DecimalField(max_digits=10, decimal_places=2)
-    active = models.BooleanField(default=True)
-
-    class Meta:
-        verbose_name = "Coupon"
-        verbose_name_plural = "Coupons"
-
-    def __str__(self):
-        return self.code
-
-
-class Cart(BaseModel):
-    buyer = models.ForeignKey(User, on_delete=models.CASCADE)
-    quantity = models.PositiveIntegerField()
-    listings = models.ManyToManyField(Listing)
-
-    class Meta:
-        verbose_name = "Order"
-        verbose_name_plural = "Orders"
-
-    def __str__(self):
-        return f"{self.buyer.username}'s cart"
-
-
 class Favorite(BaseModel):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    listing = models.ForeignKey(Listing, on_delete=models.CASCADE)
+    listing = models.ForeignKey('listings.Listing', on_delete=models.CASCADE)
 
     class Meta:
         verbose_name = "Favorite"
@@ -112,7 +56,7 @@ class Favorite(BaseModel):
 
 class Review(BaseModel):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    listing = models.ForeignKey(Listing, on_delete=models.CASCADE)
+    listing = models.ForeignKey('listings.Listing', on_delete=models.CASCADE)
     rating = models.PositiveSmallIntegerField()
     comment = models.TextField()
 
@@ -122,21 +66,6 @@ class Review(BaseModel):
 
     def __str__(self):
         return f"{self.user.username}'s review of {self.listing.title}"
-
-
-class Transaction(BaseModel):
-    buyer = models.ForeignKey(User, on_delete=models.CASCADE, related_name="buyer")
-    seller = models.ForeignKey(User, on_delete=models.CASCADE, related_name="seller")
-    listing = models.ForeignKey(Listing, on_delete=models.CASCADE)
-    quantity = models.PositiveIntegerField()
-    total = models.DecimalField(max_digits=10, decimal_places=2)
-
-    class Meta:
-        verbose_name = "Transaction"
-        verbose_name_plural = "Transactions"
-
-    def __str__(self):
-        return f"{self.buyer.username} bought {self.quantity} of {self.listing.title} from {self.seller.username}"
 
 
 class Message(BaseModel):
